actor created events now available to use whenever an actor is made. must use their type id and actorid so that projectiles and melee units can be distinguished from one another.

+++ include/GameLogic.hpp
@@ -136,7 +136,7 @@ public:
 
   const shared_ptr<TowerInterface> getTowerPlaced(int row, int col);
   const unordered_map<int, shared_ptr<TowerInterface>>& getTowersPlaced();
-  const unordered_map<int,shared_ptr<MeleeUnit>>& getSpawnedEnemyUnits();
+  const unordered_map<long long,shared_ptr<MeleeUnit>>& getSpawnedEnemyUnits();
 
   shared_ptr<WaveManager> getWaveManager(){return waveManager;};
