+++ src/WaveManager.cpp
@@ -232,4 +232,4 @@ map<int,vector<WaveManager::intPair>> WaveManager::getNormalizedDistanceMap(map<
 
 queue<shared_ptr<MeleeUnit>> WaveManager::getNextWave() {return currentWave;}
 
-unordered_map<int,shared_ptr<MeleeUnit>>& WaveManager::getSpawnedEnemyUnits(){return spawnedCurrentWave;}
+unordered_map<long long,shared_ptr<MeleeUnit>>& WaveManager::getSpawnedEnemyUnits(){return spawnedCurrentWave;}
